youtubeapi/serializer/youtube_api_key_serializer.py
from rest_framework import serializers
from youtubeapi.models import YoutubeAPIKey


class YoutubeAPIKeySerializer(serializers.ModelSerializer):

    class Meta:
        model = YoutubeAPIKey
        fields = ["key"]
